airflow_pyspark/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark
delta-spark==2.4.0
apache-airflow-providers-apache-spark
apache-airflow-providers-common-compat
boto3
google-cloud-storage
nbformat
pyarrow
findspark